requirements.txt
# Core numerical libraries (compatible versions)
numpy==1.21.6  # Compatible with both quadprog and modern libraries
scipy==1.7.3
matplotlib==3.5.3
pandas==1.3.5

# Optimization libraries
casadi==3.5.5
cvxpy==1.2.3
quadprog==0.1.11  # Latest version that works with numpy 1.21
scikit-learn==1.0.2

# Trajectory planning
trajectory-planning-helpers==0.76

# Configuration and utilities
pyyaml==6.0
configparser==5.3.0
tqdm==4.65.0

# Optional development tools
jupyter==1.0.0
ipykernel==6.19.2
pytest==7.2.0
black==22.12.0
flake8==6.0.0
